+++ src/party_menu.c
@@ -6461,7 +6461,7 @@ static void SwapFusionMonMoves(struct Pokemon *mon, const u16 moveTable[][2], u3
         {
             if (move == moveTable[j][oldMoveIndex])
             {
-                u8 maxPP = CalculateMaxPP(moveTable[j][newMoveIndex]);
+                u32 maxPP = CalculateMaxPP(moveTable[j][newMoveIndex]);
                 SetMonData(mon, MON_DATA_MOVE1 + i, &moveTable[j][newMoveIndex]);
                 SetMonData(mon, MON_DATA_PP1 + i, &maxPP);
             }
